federation/web/requirements.txt
flask==3.0.3
Flask-OIDC==2.2.0
python-dotenv==1.0.1
boto3==1.34.145
PyJWT==2.8.0
